Move StagingContext to dotty.tools.dotc.staging
